+++ ietfparse/compat/__init__.py
@@ -1,0 +1,9 @@
+"""
+Python version compatibility package.
+
+This package contains modules that isolate the rest of the code
+base from differences between Python versions.  I know that I could
+have use ``six`` but I actually dislike the ``sys.path`` hackery
+that it uses and would rather simply write what I need.
+
+"""