+++ docs/changelog.rst
@@ -1,6 +1,10 @@
 Changelog
 ---------
 
+* Next Release
+
+  - Added ``algorithms.rewrite_url``
+
 * 1.0.0 (21-Sep-2014)
 
   - Initial implementation containing the following functionality: